Contract tests wip
